+++ Adjust/ADJEvent.m
@@ -10,11 +10,12 @@
 #import "ADJAdjustFactory.h"
 #import "ADJUtil.h"
 
-#pragma mark -
 @interface ADJEvent()
+
 @property (nonatomic, weak) id<ADJLogger> logger;
-@property (nonatomic, strong) NSMutableDictionary* callbackMutableParameters;
-@property (nonatomic, strong) NSMutableDictionary* partnerMutableParameters;
+@property (nonatomic, strong) NSMutableDictionary *callbackMutableParameters;
+@property (nonatomic, strong) NSMutableDictionary *partnerMutableParameters;
+
 @end
 
 @implementation ADJEvent
@@ -23,86 +24,109 @@
     return [[ADJEvent alloc] initWithEventToken:eventToken];
 }
 
-- (id) initWithEventToken:(NSString *)eventToken {
+- (id)initWithEventToken:(NSString *)eventToken {
     self = [super init];
-    if (self == nil) return nil;
+    if (self == nil) {
+        return nil;
+    }
 
     self.logger = ADJAdjustFactory.logger;
 
-    if (![self checkEventToken:eventToken]) return self;
+    if (![self checkEventToken:eventToken]) {
+        return self;
+    }
 
-    _eventToken = eventToken;
+    _eventToken = [eventToken copy];
 
     return self;
 }
 
-- (void) addCallbackParameter:(NSString *)key
-                        value:(NSString *)value
-{
-    if (![ADJUtil isValidParameter:key
-                     attributeType:@"key"
-                     parameterName:@"Callback"]) return;
+- (void)addCallbackParameter:(NSString *)key value:(NSString *)value {
+    @synchronized (self) {
+        NSString *immutableKey = [key copy];
+        NSString *immutableValue = [value copy];
 
-    if (![ADJUtil isValidParameter:value
-                     attributeType:@"value"
-                     parameterName:@"Callback"]) return;
+        if (![ADJUtil isValidParameter:immutableKey
+                         attributeType:@"key"
+                         parameterName:@"Callback"]) {
+            return;
+        }
+        if (![ADJUtil isValidParameter:immutableValue
+                         attributeType:@"value"
+                         parameterName:@"Callback"]) {
+            return;
+        }
 
-    if (self.callbackMutableParameters == nil) {
-        self.callbackMutableParameters = [[NSMutableDictionary alloc] init];
+        if (self.callbackMutableParameters == nil) {
+            self.callbackMutableParameters = [[NSMutableDictionary alloc] init];
+        }
+        if ([self.callbackMutableParameters objectForKey:immutableKey]) {
+            [self.logger warn:@"Callback parameter key %@ was overwritten", immutableKey];
+        }
+        [self.callbackMutableParameters setObject:immutableValue forKey:immutableKey];
     }
-
-    if ([self.callbackMutableParameters objectForKey:key]) {
-        [self.logger warn:@"key %@ was overwritten", key];
-    }
-
-    [self.callbackMutableParameters setObject:value forKey:key];
 }
 
-- (void) addPartnerParameter:(NSString *)key
-                       value:(NSString *)value {
+- (void)addPartnerParameter:(NSString *)key value:(NSString *)value {
+    @synchronized (self) {
+        NSString *immutableKey = [key copy];
+        NSString *immutableValue = [value copy];
 
-    if (![ADJUtil isValidParameter:key
-                     attributeType:@"key"
-                     parameterName:@"Partner"]) return;
+        if (![ADJUtil isValidParameter:immutableKey
+                         attributeType:@"key"
+                         parameterName:@"Partner"]) {
+            return;
+        }
+        if (![ADJUtil isValidParameter:immutableValue
+                         attributeType:@"value"
+                         parameterName:@"Partner"]) {
+            return;
+        }
 
-    if (![ADJUtil isValidParameter:value
-                     attributeType:@"value"
-                     parameterName:@"Partner"]) return;
-
-    if (self.partnerMutableParameters == nil) {
-        self.partnerMutableParameters = [[NSMutableDictionary alloc] init];
+        if (self.partnerMutableParameters == nil) {
+            self.partnerMutableParameters = [[NSMutableDictionary alloc] init];
+        }
+        if ([self.partnerMutableParameters objectForKey:immutableKey]) {
+            [self.logger warn:@"Partner parameter key %@ was overwritten", immutableKey];
+        }
+        [self.partnerMutableParameters setObject:immutableValue forKey:immutableKey];
     }
-
-    if ([self.partnerMutableParameters objectForKey:key]) {
-        [self.logger warn:@"key %@ was overwritten", key];
-    }
-
-    [self.partnerMutableParameters setObject:value forKey:key];
 }
 
-- (void) setRevenue:(double) amount currency:(NSString *)currency{
-    NSNumber * revenue = [NSNumber numberWithDouble:amount];
-
-    if (![self checkRevenue:revenue currency:currency]) return;
+- (void)setRevenue:(double)amount currency:(NSString *)currency {
+    NSNumber *revenue = [NSNumber numberWithDouble:amount];
+    if (![self checkRevenue:revenue currency:currency]) {
+        return;
+    }
 
     _revenue = revenue;
-    _currency = currency;
+    @synchronized (self) {
+        _currency = [currency copy];
+    }
 }
 
-- (void) setTransactionId:(NSString *)transactionId {
-    _transactionId = transactionId;
+- (void)setTransactionId:(NSString *)transactionId {
+    @synchronized (self) {
+        _transactionId = [transactionId copy];
+    }
 }
 
 - (void)setCallbackId:(NSString *)callbackId {
-    _callbackId = callbackId;
+    @synchronized (self) {
+        _callbackId = [callbackId copy];
+    }
 }
 
-- (NSDictionary *) callbackParameters {
-    return (NSDictionary *) self.callbackMutableParameters;
+- (NSDictionary *)callbackParameters {
+    @synchronized (self) {
+        return (NSDictionary *)self.callbackMutableParameters;
+    }
 }
 
-- (NSDictionary *) partnerParameters {
-    return (NSDictionary *) self.partnerMutableParameters;
+- (NSDictionary *)partnerParameters {
+    @synchronized (self) {
+        return (NSDictionary *)self.partnerMutableParameters;
+    }
 }
 
 - (BOOL)checkEventToken:(NSString *)eventToken {
@@ -117,21 +141,17 @@
     return YES;
 }
 
-- (BOOL) checkRevenue:(NSNumber*) revenue
-             currency:(NSString*) currency
-{
+- (BOOL)checkRevenue:(NSNumber *)revenue currency:(NSString *)currency {
     if (![ADJUtil isNull:revenue]) {
         double amount =  [revenue doubleValue];
         if (amount < 0.0) {
             [self.logger error:@"Invalid amount %.5f", amount];
             return NO;
         }
-
         if ([ADJUtil isNull:currency]) {
             [self.logger error:@"Currency must be set with revenue"];
             return NO;
         }
-
         if ([currency isEqualToString:@""]) {
             [self.logger error:@"Currency is empty"];
             return NO;
@@ -146,12 +166,14 @@
     return YES;
 }
 
-- (BOOL) isValid {
+- (BOOL)isValid {
     return self.eventToken != nil;
 }
 
-- (void) setReceipt:(NSData *)receipt transactionId:(NSString *)transactionId {
-    if (![self checkReceipt:receipt transactionId:transactionId]) return;
+- (void)setReceipt:(NSData *)receipt transactionId:(NSString *)transactionId {
+    if (![self checkReceipt:receipt transactionId:transactionId]) {
+        return;
+    }
 
     if ([ADJUtil isNull:receipt] || [receipt length] == 0) {
         _emptyReceipt = YES;
@@ -160,7 +182,7 @@
     _transactionId = transactionId;
 }
 
-- (BOOL) checkReceipt:(NSData *)receipt transactionId:(NSString *)transactionId {
+- (BOOL)checkReceipt:(NSData *)receipt transactionId:(NSString *)transactionId {
     if ([ADJUtil isNotNull:receipt] && [ADJUtil isNull:transactionId]) {
         [self.logger error:@"Missing transactionId"];
         return NO;
@@ -168,9 +190,9 @@
     return YES;
 }
 
--(id)copyWithZone:(NSZone *)zone
-{
-    ADJEvent* copy = [[[self class] allocWithZone:zone] init];
+- (id)copyWithZone:(NSZone *)zone {
+    ADJEvent *copy = [[[self class] allocWithZone:zone] init];
+
     if (copy) {
         copy->_eventToken = [self.eventToken copyWithZone:zone];
         copy->_revenue = [self.revenue copyWithZone:zone];
@@ -181,6 +203,7 @@
         copy->_receipt = [self.receipt copyWithZone:zone];
         copy->_emptyReceipt = self.emptyReceipt;
     }
+
     return copy;
 }