+++ AdjustBridge/AdjustBridgeRegister.m
@@ -221,7 +221,7 @@ static NSString * fbAppIdStatic = nil;
                 if (this.sdkPrefix) {
                     return this.sdkPrefix;
                 } else {
-                    return 'web-bridge4.21.2';
+                    return 'web-bridge4.21.3';
                 }
             },
             setTestOptions: function(testOptions) {